Some Architecture fixes, api WIP

+++ app/src/main/java/com/example/loansapp/presentation/loanDetailsScreen/LoanDetailViewModel.kt
@@ -8,8 +8,8 @@ import com.example.loansapp.data.ChosenLoanRepository
 class LoanDetailViewModel(private var repository: ChosenLoanRepository, private var login:String): ViewModel() {
     private var loanName = MutableLiveData<String>(repository.getChosenLoan()?.name ?: "")
     private var loanId = MutableLiveData<Long>(repository.getChosenLoan()?.id ?: 0)
-    private var loanAmount = MutableLiveData<Double>(repository.getChosenLoan()?.amount ?: 0.0)
-    private var loanAmountLeft = MutableLiveData<Double>(repository.getChosenLoan()?.amount_left ?: 0.0)
+    private var loanAmount = MutableLiveData<Int>(repository.getChosenLoan()?.amount ?: 0)
+    private var loanAmountLeft = MutableLiveData<Int>(repository.getChosenLoan()?.amount_left ?: 0)
     private var loanRate = MutableLiveData<Double>(repository.getChosenLoan()?.rate ?: 0.0)
     private var loanStartDate = MutableLiveData<String>(repository.getChosenLoan()?.start_date ?: "")
     private var loanExpDate = MutableLiveData<String>(repository.getChosenLoan()?.expiration_date ?: "")
@@ -20,10 +20,10 @@ class LoanDetailViewModel(private var repository: ChosenLoanRepository, private 
     fun getId(): LiveData<Long> {
         return loanId
     }
-    fun getAmount(): LiveData<Double> {
+    fun getAmount(): LiveData<Int> {
         return loanAmount
     }
-    fun getAmountLeft(): LiveData<Double> {
+    fun getAmountLeft(): LiveData<Int> {
         return loanAmountLeft
     }
     fun getRate(): LiveData<Double> {
